refactor: Extract file type detection logic into separate module
